+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - General:
   - Search button in text and regex searches (you can still press "enter")
 - External scripts:
-  - Script definitions supports `values` shown in a combo ([sample](ext-search-samples/cyggrep.jloga.json))
+  - Script definitions supports `values` shown in a combo or checkbox ([sample](ext-search-samples/cyggrep.jloga.json))
 - Bug fixes:
   - Toggle buttons in text and regex
   - Fix for "Key too long" error